+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.4-pre
-*
+* Add rsync protocol support
 
 ### 0.0.3
 * Add SCP protocol support